valhalla/requirements.txt
asgiref==3.4.1
Django==4.0.1
sqlparse==0.4.2
django-model-utils==4.1.1
django-filter==2.4.0
django-crispy-forms==1.10.0
django-celery-beat==2.3.0
django-cors-headers==3.8.0
python-dotenv==0.15.0
elasticsearch==7.10.1
django-rest-swagger==2.2.0
requests==2.25.1
grequests==0.6.0
redis==3.5.3
djangorestframework==3.12.2
djangorestframework-datatables==0.5.2
django-import-export==2.5.0
django-admin-rangefilter==0.7.0
django-environ==0.4.5
Pillow==8.3.1
instaloader==4.8.1
gunicorn==20.1.0
drf-yasg==1.20.0
psycopg2-binary==2.9.3
drf-excel==2.1.0
sentry-sdk==1.9.8
pandas==1.5.1